# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: Deep Scrubber API
  version: 0.1.0
  description: Authorized personal-data discovery, removal, verification and monitoring. See docs/api.md for semantics.
servers:
- url: http://localhost:4000
security:
- session: []
  csrf: []
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: ds_session
    csrf:
      type: apiKey
      in: header
      name: x-csrf-token
      description: Required on non-GET requests.
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
                  details:
                    type: object
paths:
  /api/abuse-reports:
    post:
      tags:
      - misc
      operationId: post_abuse_reports
      responses:
        '200':
          description: OK
        4XX:
          $ref: '#/components/responses/Error'
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                category:
                  type: string
                description:
                  type: string
                contact:
                  type: string
  /api/account:
    delete:
      tags:
      - misc
      operationId: delete_account
      responses:
        '200':
          description: OK
        4XX:
          $ref: '#/components/responses/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
                confirm:
                  type: string
  /api/account/export:
    get:
      tags:
      - misc
      operationId: get_account_export
      responses:
        '200':
          description: OK
        4XX:
          $ref: '#/components/responses/Error'
...
  /api/billing/change-plan:
    post:
      tags:
      - misc
      operationId: post_billing_change_plan
      responses:
        '200':
          description: OK
        4XX:
          $ref: '#/components/responses/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                plan:
                  type: string
  /api/billing/plans:
    get:
      tags:
      - misc
      operationId: get_billing_plans
      responses:
        '200':
          description: OK
        4XX:
          $ref: '#/components/responses/Error'
      security: []
  /api/dashboard:
    get:
      tags:
      - misc
      operationId: get_dashboard
      responses:
        '200':
          description: OK
        4XX:
          $ref: '#/components/responses/Error'
  /api/explanations:
    get:
      tags:
      - misc
      operationId: get_explanations
      responses:
        '200':
          description: OK
        4XX:
          $ref: '#/components/responses/Error'
      security: []
...
  /api/inbound-email:
    post:
      tags:
      - misc
      operationId: post_inbound_email
      responses:
        '200':
          description: OK
        4XX:
          $ref: '#/components/responses/Error'
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                to:
                  type: string
                from:
                  type: string
                subject:
                  type: string
                text:
                  type: string
  /api/jurisdictions:
    get:
      tags:
      - misc
      operationId: get_jurisdictions
      responses:
        '200':
          description: OK
        4XX:
          $ref: '#/components/responses/Error'
      security: []
  /api/monitoring:
    get:
      tags:
      - misc
      operationId: get_monitoring
      responses:
        '200':
          description: OK
        4XX:
          $ref: '#/components/responses/Error'
    post:
      tags:
      - misc
      operationId: post_monitoring
      responses:
        '200':
          description: OK
        4XX:
          $ref: '#/components/responses/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                recordId:
                  type: string
                intervalDays:
                  type: integer
                enabled:
                  type: boolean
  /api/monitoring/{id}:
    delete:
      tags:
      - misc
      operationId: delete_monitoring_id
      responses:
        '200':
          description: OK
        4XX:
          $ref: '#/components/responses/Error'
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
  /api/notifications:
    get:
      tags:
      - misc
      operationId: get_notifications
      responses:
        '200':
          description: OK
        4XX:
          $ref: '#/components/responses/Error'
  /api/notifications/read-all:
    post:
      tags:
      - misc
      operationId: post_notifications_read_all
      responses:
        '200':
          description: OK
        4XX:
          $ref: '#/components/responses/Error'
  /api/notifications/{id}/read:
    post:
      tags:
      - misc
      operationId: post_notifications_id_read
      responses:
        '200':
          description: OK
        4XX:
          $ref: '#/components/responses/Error'
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
...
  /api/push/subscribe:
    post:
      tags:
      - misc
      operationId: post_push_subscribe
      responses:
        '200':
          description: OK
        4XX:
          $ref: '#/components/responses/Error'
...
  /api/reports/export.csv:
    get:
      tags:
      - misc
      operationId: get_reports_export_csv
      responses:
        '200':
          description: OK
        4XX:
          $ref: '#/components/responses/Error'
  /api/reports/export.pdf:
    get:
      tags:
      - misc
      operationId: get_reports_export_pdf
      responses:
        '200':
          description: OK
        4XX:
          $ref: '#/components/responses/Error'
  /api/reports/summary:
    get:
      tags:
      - misc
      operationId: get_reports_summary
      responses:
        '200':
          description: OK
        4XX:
          $ref: '#/components/responses/Error'
...
  /api/score/history:
    get:
      tags:
      - misc
      operationId: get_score_history
      responses:
        '200':
          description: OK
        4XX:
          $ref: '#/components/responses/Error'
  /api/settings:
    get:
      tags:
      - misc
      operationId: get_settings
      responses:
        '200':
          description: OK
        4XX:
          $ref: '#/components/responses/Error'
    put:
      tags:
      - misc
      operationId: put_settings
      responses:
        '200':
          description: OK
        4XX:
          $ref: '#/components/responses/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                notifyEmail:
                  type: boolean
                notifyPush:
                  type: boolean
                notifyInApp:
                  type: boolean
                recordRetentionDays:
                  type: integer
  /api/sources:
    get:
      tags:
      - misc
      operationId: get_sources
      responses:
        '200':
          description: OK
        4XX:
          $ref: '#/components/responses/Error'
  /api/sources/{sourceId}/preference:
    put:
      tags:
      - misc
      operationId: put_sources_sourceId_preference
      responses:
        '200':
          description: OK
        4XX:
          $ref: '#/components/responses/Error'
      parameters:
      - name: sourceId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                approvalMode:
                  type: string